Move test utils inside the input_control_vis (#55169)
* Move test utils to an appropriate folder

* Change imports
