+++ src/legacy/core_plugins/input_control_vis/public/test_utils/get_index_patterns_mock.ts
