+++ src/legacy/core_plugins/input_control_vis/public/test_utils/update_component.ts
